+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Version 9.1.0-dev
 - allowed setting base url for Nette\Application\LinkGenerator from Nette DIC extension
+- added argument --bootstrap for mytester.php
 
 Version 9.0.2
 - fixed parameter $config of ContainerFactory::create()/TCompiledContainer::refreshContainer() not working